`afk-longest` rewritten and merged
